moved showHelp test to testUtilsCommon.js
