+++ tests/unit_tests/commands/android/testIndex.js
@@ -5,134 +5,6 @@ const path = require('path');
 const os = require('os');
 import { ABI } from '../../../../src/commands/android/constants';
 
-describe('test showHelp', function() {
-  beforeEach(() => {
-    mockery.enable({useCleanCache: true, warnOnReplace: false, warnOnUnregistered: false});
-  });
-
-  afterEach(() => {
-    mockery.deregisterAll();
-    mockery.resetCache();
-    mockery.disable();
-  });
-
-  it('shows error for unknown options and no unknown subcommand', () => {
-    const consoleOutput = [];
-    mockery.registerMock(
-      '../../../logger',
-      class {
-        static log(...msgs) {
-          consoleOutput.push(...msgs);
-        }
-      }
-    );
-
-    const colorFn = (arg) => arg;
-    mockery.registerMock('ansi-colors', {
-      green: colorFn,
-      yellow: colorFn,
-      magenta: colorFn,
-      cyan: colorFn,
-      red: colorFn,
-      grey: colorFn
-    });
-
-    const {showHelp} = require('../../../../src/commands/android/utils/common');
-    showHelp(['random-option']);
-
-    const output = consoleOutput.toString();
-    assert.strictEqual(output.includes('unknown option(s) passed: random-option'), true);
-    assert.strictEqual(output.includes('unknown subcommand passed:'), false);
-  });
-
-  it('shows error for no unknown options and unknown subcommand', () => {
-    const consoleOutput = [];
-    mockery.registerMock(
-      '../../../logger',
-      class {
-        static log(...msgs) {
-          consoleOutput.push(...msgs);
-        }
-      }
-    );
-
-    const colorFn = (arg) => arg;
-    mockery.registerMock('ansi-colors', {
-      green: colorFn,
-      yellow: colorFn,
-      magenta: colorFn,
-      cyan: colorFn,
-      red: colorFn,
-      grey: colorFn
-    });
-
-    const {showHelp} = require('../../../../src/commands/android/utils/common');
-    showHelp([], 'random-command');
-
-    const output = consoleOutput.toString();
-    assert.strictEqual(output.includes('unknown option(s) passed:'), false);
-    assert.strictEqual(output.includes('unknown subcommand passed: random-command'), true);
-  });
-
-  it('shows error for unknown options and unknown subcommand', () => {
-    const consoleOutput = [];
-    mockery.registerMock(
-      '../../../logger',
-      class {
-        static log(...msgs) {
-          consoleOutput.push(...msgs);
-        }
-      }
-    );
-
-    const colorFn = (arg) => arg;
-    mockery.registerMock('ansi-colors', {
-      green: colorFn,
-      yellow: colorFn,
-      magenta: colorFn,
-      cyan: colorFn,
-      red: colorFn,
-      grey: colorFn
-    });
-
-    const {showHelp} = require('../../../../src/commands/android/utils/common');
-    showHelp(['random-option'], 'random-command');
-
-    const output = consoleOutput.toString();
-    assert.strictEqual(output.includes('unknown option(s) passed:'), false);
-    assert.strictEqual(output.includes('unknown subcommand passed: random-command'), true);
-  });
-
-  it('does not shows error for no unknown options and no unknown subcommand', () => {
-    const consoleOutput = [];
-    mockery.registerMock(
-      '../../../logger',
-      class {
-        static log(...msgs) {
-          consoleOutput.push(...msgs);
-        }
-      }
-    );
-
-    const colorFn = (arg) => arg;
-    mockery.registerMock('ansi-colors', {
-      green: colorFn,
-      yellow: colorFn,
-      magenta: colorFn,
-      cyan: colorFn,
-      red: colorFn,
-      grey: colorFn
-    });
-
-    const {showHelp} = require('../../../../src/commands/android/utils/common');
-    showHelp([]);
-
-    const output = consoleOutput.toString();
-    assert.strictEqual(output.includes('unknown option(s) passed:'), false);
-    assert.strictEqual(output.includes('unknown subcommand passed:'), false);
-  });
-});
-
 describe('test checkJavaInstallation', function() {
   beforeEach(() => {
     mockery.enable({useCleanCache: true, warnOnReplace: false, warnOnUnregistered: false});
